requirements.txt
# ===================================================================
# 🚀 RAG Enhanced - Requirements File
# ===================================================================
# 
# Este arquivo contém todas as dependências necessárias para o
# projeto RAG Enhanced, incluindo o sistema principal, framework
# de testes, interfaces e utilitários.
#
# Instalação:
#   pip install -r requirements.txt
#
# Para desenvolvimento:
#   pip install -r requirements.txt -r requirements-dev.txt
# ===================================================================

# ===================================================================
# 📦 DEPENDÊNCIAS PRINCIPAIS
# ===================================================================

# Google Cloud e AI
google-cloud-storage>=2.10.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Interface de usuário
gradio>=4.0.0

# Processamento de dados
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Validação e parsing
pydantic>=2.0.0
PyYAML>=6.0.1
jsonschema>=4.17.0

# Utilitários de sistema
pathlib2>=2.3.7; python_version < "3.4"
dataclasses>=0.8; python_version < "3.7"
typing-extensions>=4.0.0

# ===================================================================
# 🔧 DEPENDÊNCIAS DE DESENVOLVIMENTO E TESTES
# ===================================================================

# Framework de testes
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Mocks e simulação
responses>=0.23.0
freezegun>=1.2.0

# Análise de código
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# ===================================================================
# 📊 ANÁLISE E MONITORAMENTO
# ===================================================================

# Monitoramento de sistema
psutil>=5.9.0

# Logging e debugging
colorama>=0.4.6
rich>=13.0.0
loguru>=0.7.0

# Progress bars
tqdm>=4.65.0

# ===================================================================
# 🌐 REDE E HTTP
# ===================================================================

# Cliente HTTP
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# ===================================================================
# 📁 MANIPULAÇÃO DE ARQUIVOS
# ===================================================================

# Compressão e arquivos
zipfile38>=0.0.3; python_version < "3.8"
chardet>=5.1.0
python-magic>=0.4.27

# ===================================================================
# 🎨 INTERFACE E CLI
# ===================================================================

# CLI frameworks
click>=8.1.0
typer>=0.9.0
argparse>=1.4.0; python_version < "3.2"

# ===================================================================
# 🔒 SEGURANÇA E AUTENTICAÇÃO
# ===================================================================

# Criptografia e segurança
cryptography>=40.0.0
python-dotenv>=1.0.0

# ===================================================================
# 📈 ANÁLISE DE DADOS (OPCIONAL)
# ===================================================================

# Visualização (opcional para relatórios)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# ===================================================================
# 🚀 PERFORMANCE E CONCORRÊNCIA
# ===================================================================

# Async e concorrência
asyncio>=3.4.3; python_version < "3.7"
concurrent-futures>=3.1.1; python_version < "3.2"
aiofiles>=23.0.0

# ===================================================================
# 🧪 DEPENDÊNCIAS ESPECÍFICAS DO FRAMEWORK DE TESTES
# ===================================================================

# Geração de dados de teste
Faker>=18.0.0
factory-boy>=3.2.0

# Validação avançada
cerberus>=1.3.4
marshmallow>=3.19.0

# ===================================================================
# 📝 DOCUMENTAÇÃO
# ===================================================================

# Geração de documentação
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
mkdocs>=1.4.0
mkdocs-material>=9.0.0

# ===================================================================
# 🔄 COMPATIBILIDADE
# ===================================================================

# Compatibilidade com versões antigas do Python
importlib-metadata>=6.0.0; python_version < "3.8"
importlib-resources>=5.0.0; python_version < "3.9"
backports.zoneinfo>=0.2.1; python_version < "3.9"

# ===================================================================
# 📦 EMPACOTAMENTO E DISTRIBUIÇÃO
# ===================================================================

# Build e distribuição
setuptools>=65.0.0
wheel>=0.40.0
build>=0.10.0
twine>=4.0.0

# ===================================================================
# 🎯 DEPENDÊNCIAS ESPECÍFICAS POR FUNCIONALIDADE
# ===================================================================

# Para análise de código
ast-decompiler>=0.7.0
astpretty>=3.0.0

# Para processamento de texto
nltk>=3.8.0
spacy>=3.5.0

# Para manipulação de imagens (multimodal)
Pillow>=9.5.0
opencv-python>=4.7.0

# Para processamento de documentos
python-docx>=0.8.11
PyPDF2>=3.0.0
markdown>=3.4.0

# ===================================================================
# 🌍 INTERNACIONALIZAÇÃO
# ===================================================================

# Suporte a idiomas
babel>=2.12.0
python-i18n>=0.3.9

# ===================================================================
# 📊 MÉTRICAS E TELEMETRIA (OPCIONAL)
# ===================================================================

# Métricas (opcional)
prometheus-client>=0.16.0
statsd>=4.0.0

# ===================================================================
# 🔧 UTILITÁRIOS ADICIONAIS
# ===================================================================

# Utilitários diversos
python-dateutil>=2.8.0
pytz>=2023.3
humanize>=4.6.0
slugify>=0.0.1

# Validação de email e URLs
email-validator>=2.0.0
validators>=0.20.0

# ===================================================================
# 🎮 DEPENDÊNCIAS DE DESENVOLVIMENTO LOCAL
# ===================================================================

# Jupyter para desenvolvimento
jupyter>=1.0.0
ipython>=8.0.0
notebook>=6.5.0

# Ferramentas de profiling
memory-profiler>=0.60.0
line-profiler>=4.0.0

# ===================================================================
# 🔍 LINTING E FORMATAÇÃO
# ===================================================================

# Linting adicional
pylint>=2.17.0
bandit>=1.7.0
safety>=2.3.0

# Formatação de código
autopep8>=2.0.0
yapf>=0.32.0

# ===================================================================
# 📋 NOTAS DE INSTALAÇÃO
# ===================================================================
#
# 1. INSTALAÇÃO BÁSICA:
#    pip install -r requirements.txt
#
# 2. INSTALAÇÃO PARA DESENVOLVIMENTO:
#    pip install -r requirements.txt
#    pip install -r requirements-dev.txt
#
# 3. INSTALAÇÃO MÍNIMA (apenas funcionalidades core):
#    pip install google-cloud-storage google-generativeai vertexai gradio pandas pyyaml
#
# 4. PROBLEMAS COMUNS:
#    - Se houver conflitos de versão, use: pip install --upgrade --force-reinstall
#    - Para ambientes virtuais: python -m venv venv && source venv/bin/activate
#    - No Windows: python -m venv venv && venv\Scripts\activate
#
# 5. VERIFICAÇÃO:
#    python -c "import google.cloud.storage, vertexai, gradio; print('✅ Dependências OK')"
#
# ===================================================================